requirements.txt
cloudinary==1.39.1
Django==5.0.3
django_allauth==0.61.1
django_environ==0.11.2
python-dotenv==1.0.1
selenium==4.19.0
django-extensions==3.2.3
pyodbc==5.1.0
pymssql==2.3.0
mssql-django==1.4.2
locust==2.24.1
django-storages==1.14.2
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
django-cron==0.6.0
drf-yasg==1.21.7





